pyproject.toml
[project]
name = "mcp-servers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi-mcp>=0.3.4",
    "fastapi[standard]>=0.115.12",
    "pydantic-settings>=2.9.1",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "reverse-geocoder>=1.5.1",
    "typer>=0.16.0",
]
